refactoring: creating BookStoreApplication package, creating new test classes, moving test, testNavigationToBookStorePage is added
