+++ src/test/java/pages/BookStoreApplication/LoginPage.java
@@ -1,9 +1,7 @@
 package pages.BookStoreApplication;
 
-import com.fasterxml.jackson.databind.ser.Serializers;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.devtools.v85.profiler.model.Profile;
 import org.openqa.selenium.support.FindBy;
 import pages.BasePage;
 
@@ -43,6 +41,7 @@ public class LoginPage extends BookStorePage {
     /*
     при логине с невалидными данными мы остаемся на LoginPage
     при логине с валидными данными мы переходим или на ProfilePage или на BookStorePage
+    и мне не нравится этот метод, если что :)
      */
     public  <T extends BookStorePage>T clickLoginButton() {
         click(loginButton);
@@ -59,6 +58,7 @@ public class LoginPage extends BookStorePage {
     }
 
     public RegisterPage clickNewUserButton() {
+        wait20ElementToBeVisible(newUserButton);
         click(newUserButton);
 
         return new RegisterPage(getDriver());
