+++ src/test/java/tests/BookStoreApplicationTest/BookStoreAPITest.java
@@ -1,0 +1,7 @@
+package tests.BookStoreApplicationTest;
+
+import base.BaseTest;
+
+public class BookStoreAPITest extends BaseTest {
+
+}
