+++ src/test/java/tests/BookStoreApplicationTest/ProfileTest.java
@@ -1,0 +1,7 @@
+package tests.BookStoreApplicationTest;
+
+import base.BaseTest;
+
+public class ProfileTest extends BaseTest {
+
+}
